requirements.txt
numpy>=1.26
torch>=2.2
gymnasium>=0.29
matplotlib>=3.8
PyYAML>=6.0